+++ core/src/io/anuke/mindustry/entities/enemies/Enemy.java
@@ -72,7 +72,7 @@ public class Enemy extends SyncEntity {
 	}
 
 	@Override
-	public void damage(int amount){
+	public void damage(float amount){
 		super.damage(amount);
 		hitTime = EnemyType.hitDuration;
 	}
