+++ core/src/io/anuke/mindustry/io/versions/Save14.java
@@ -225,7 +225,7 @@ public class Save14 extends SaveFileVersion{
         stream.writeFloat(Vars.player.x); //player x/y
         stream.writeFloat(Vars.player.y);
 
-        stream.writeInt(Vars.player.health); //player health
+        stream.writeInt((int)Vars.player.health); //player health
 
         stream.writeByte(control.upgrades().getWeapons().size - 1); //amount of weapons
 
@@ -267,7 +267,7 @@ public class Save14 extends SaveFileVersion{
             stream.writeFloat(enemy.x); //x
             stream.writeFloat(enemy.y); //y
             stream.writeByte(enemy.tier); //tier
-            stream.writeShort(enemy.health); //health
+            stream.writeShort((short)enemy.health); //health
         }
 
         //--MAP DATA--
